requirements.txt
bidict==0.23.1
blinker==1.9.0
cachelib==0.13.0
click==8.1.7
colorama==0.4.6
Flask==3.1.0
Flask-Session==0.8.0
Flask-SocketIO==5.4.1
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
msgspec==0.18.6
packaging==24.2
psycopg2==2.9.10
python-dotenv==1.0.1
python-engineio==4.11.1
python-socketio==5.11.4
redis==5.2.1
simple-websocket==1.1.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
waitress==3.0.2
Werkzeug==3.1.3
wsproto==1.2.0
